Hit Detection
Originally subtracted ten everytime the asterisk hit the player but will
now only subtract it once and breaks the loop so that only one will hit
the player. However, I ran into a problem with detecting when it hits
the player because it now stops when it hits the entire row. Altered the
bool so that the true statement comes first and is easier to understand.